requirements.txt
# dependencies
beautifulsoup4
colorama
md2pdf
python-dotenv
pyyaml
uvicorn
pydantic
pydantic[email]
fastapi
python-multipart
markdown
langchain
langchain_community
langchain-openai
langchain-ollama
langgraph
tiktoken
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru
PyJWT
passlib

# uncomment for testing
# pytest
# pytest-asyncio
